# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hospital Water Tank Monitoring API
  description: Backend REST API & Realtime Monitoring Specification for Integrated Hospital Water Tank Monitoring Research Project (RSA UGM).
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local Development Server
paths:
...
  /api/users:
    get:
      summary: List all users (Admin only)
      description: Retrieve list of system operators and admins.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin role required)
    post:
      summary: Create user (Admin only)
      description: Register a new operator, viewer, or admin account.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserInput'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Pengguna baru berhasil dibuat"
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin role required)
        '409':
          description: Username already in use
  /api/users/{id}:
    patch:
      summary: Update user role / active status / password (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserInput'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Pengguna berhasil diperbarui"
                  data:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin role required)
        '404':
          description: User not found
    put:
      summary: Update user (Idempotent, Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserInput'
      responses:
        '200':
          description: User updated successfully
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: User not found
    delete:
      summary: Delete user account (Admin only)
      description: Remove a user account permanently from the system.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "User deleted successfully"
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin role required)
        '404':
          description: User not found
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          example: "admin"
        role:
          type: string
          enum: [admin, operator, viewer]
          example: "admin"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
...
    CreateUserInput:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          example: "operator_shift1"
        password:
          type: string
          minLength: 6
          example: "passwordAman123"
        role:
          type: string
          enum: [admin, operator, viewer]
          default: operator
    UpdateUserInput:
      type: object
      properties:
        role:
          type: string
          enum: [admin, operator, viewer]
        isActive:
          type: boolean
        password:
          type: string
          minLength: 6
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT Bearer token obtained from /api/auth/login
    GatewayKeyAuth:
      type: apiKey
      in: header
      name: X-Gateway-Key
      description: Pre-shared API Key header for LoRa Gateway synchronization